pyproject.toml
[tool.poetry]
name = "movies"
version = "0.1.0"
description = ""
authors = ["Javier Santacruz <javier.santacruz.lc@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.7,<3.9"
imdbpy = "^6.8"
subliminal = "^2.1.0"
tabulate = "^0.8.7"
jmespath = "^0.10.0"
pex = "^2.1.11"
shiv = "^0.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
